backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.0
python-dotenv==1.0.1
groq==1.0.0
SQLAlchemy==2.0.36
Werkzeug==3.1.3
